added circle meshes for top and bottom of cylinders

+++ main.py
@@ -3,7 +3,7 @@ from visualize3d import visualization
 
 def main():
     # visualization("short" or "long", # intruders)
-    graph = visualization("long",20)
+    graph = visualization("short",20)
 
     while(1):
         graph.update_graph()
